Syntax tree optimization and move insertion
