+++ src/compilation.rs
@@ -642,6 +642,7 @@ impl NessaContext {
 
     pub fn compile(&mut self, body: &mut Vec<NessaExpr>, args: &Vec<(String, Type)>) -> Result<(), NessaError> {
         self.compile_vars_and_infer(body, args)?;
+        self.optimize(body);
 
         Ok(())
     }
